requirements_temp.txt
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.18.1
anyio==4.12.0
asgiref==3.11.0
astroid==3.0.3
asttokens==3.0.1
attrs==25.4.0
autobahn==25.11.1
Automat==25.4.16
billiard==4.2.4
black==24.1.1
cachetools==6.2.2
cbor2==5.7.1
celery==5.3.6
certifi==2025.11.12
cffi==2.0.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
coverage==7.4.1
cryptography==42.0.2
daphne==4.0.0
decorator==5.2.1
dill==0.4.0
distro==1.9.0
Django==5.0.1
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.1
django-stubs==4.2.7
django-stubs-ext==5.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djangorestframework-stubs==3.14.5
drf-spectacular==0.27.0
executing==2.2.1
factory-boy==3.3.0
Faker==22.6.0
fastapi==0.109.0
filelock==3.20.0
flake8==7.0.0
fsspec==2025.12.0
google-ai-generativelanguage==0.4.0
google-api-core==2.28.1
google-auth==2.43.0
google-generativeai==0.3.2
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.62.3
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.2.1
hvac==2.1.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
iniconfig==2.3.0
ipdb==0.13.13
ipython==8.21.0
isort==5.13.2
jedi==0.19.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.1
Markdown==3.10
matplotlib-inline==0.2.1
mccabe==0.7.0
msgpack==1.1.2
mypy==1.8.0
mypy_extensions==1.1.0
numpy==2.3.5
openai==1.10.0
packaging==25.0
parso==0.8.5
pathspec==0.12.1
pgvector==0.2.5
pillow==12.0.0
pip==23.2.1
platformdirs==4.5.1
pluggy==1.6.0
prometheus-client==0.19.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==4.25.8
pure_eval==0.2.3
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.11.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.12.0
pyflakes==3.2.0
Pygments==2.19.2
PyJWT==2.10.1
pylint==3.0.3
pyOpenSSL==25.1.0
pyotp==2.9.0
pypng==0.20220715.0
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
pytest-django==4.7.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==8.0.1
pytz==2024.1
PyYAML==6.0.3
qrcode==7.4.2
redis==5.0.1
referencing==0.37.0
requests==2.31.0
rpds-py==0.30.0
rsa==4.9.1
sentry-sdk==1.40.0
service-identity==24.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.4
stack-data==0.6.3
starlette==0.35.1
text-unidecode==1.3
tokenizers==0.22.1
tomlkit==0.13.3
tqdm==4.67.1
traitlets==5.14.3
Twisted==25.5.0
txaio==25.12.1
typer-slim==0.20.0
types-pytz==2025.2.0.20251108
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.6.0
uvicorn==0.27.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
zope.interface==8.1.1
